Clean up mutability regarding Listing
Now, instead of using observers, Listing gets cloned everywhere and most types
that use it will store a copy and have an "update" function to bring them up to
speed with the current state of things. This opens up opportunities for cool
things like undo and changing address space objects (when we change address
space mapping).